Refactor Widgets in files
