+++ lib/constants/movie_texts.dart
@@ -1,0 +1,9 @@
+class MovieTexts {
+  static const voteAverage = '7.8';
+  static const title = 'The Super Mario Bros. Movie';
+  static const date = '2023-04-05';
+  static const genres = '[Animation, Family, Adventure, Fantasy, Comedy]';
+  static const overviewTitle = 'Overview';
+  static const overview =
+      'While working underground to fix a water main, Brooklyn plumbers—and brothers—Mario and Luigi are transported down a mysterious pipe and wander into a magical new world. But when the brothers are separated, Mario embarks on an epic quest to find Luigi.';
+}
