+++ lib/constants/ui_constants.dart
@@ -1,9 +1,7 @@
-import 'package:flutter/cupertino.dart';
-
- class UIConstants {
-     static const backdropLeftPadding = 90.0;
-     static const imagePadding = 20.0;
-     static const titleFontSize = 25.0;
-     static const textsLargePadding = 20.0;
-     static const textsShortPadding = 10.0;
+class UIConstants {
+  static const backdropLeftPadding = 90.0;
+  static const imagePadding = 20.0;
+  static const titleFontSize = 25.0;
+  static const textsLargePadding = 20.0;
+  static const textsShortPadding = 10.0;
 }
